various small code reorganizations
